// client/src/interfaces/respositoriesResponse.tsx
// Generated by https://quicktype.io

export interface RepositoriesResponse {
  viewer: Viewer;
}

export interface Viewer {
  __typename:   string;
  avatarUrl:string;
  login:string;
  repositories: RepositoriesClass;
}

export interface RepositoriesClass {
  __typename: string;
  totalCount: number;
  pageInfo:   PageInfo;
  nodes:      Node[];
}

export interface Node {
  id:string;
  __typename:  string;
  name:        string;
  description: null;
  createdAt:   string;
  updatedAt: string;
  isPrivate: boolean;
  favorite:    boolean;
  owner:       Owner;
}

export interface Owner {
  __typename: string;
  login:      string;
}

export interface PageInfo {
  __typename:  string;
  endCursor:   string;
  hasNextPage: boolean;
}